\vphantom is too fragile in ◊ and □
The height of \Diamond and \Box is really bad (it is smaller than the
symbol appears in print).  So I wanted to have something in here to fix
the height, but \vphantom does not play well with some packages like
"graphs" from tikz.  I tried to replace \vphantom with a zero-width
\rule, but this has the same problem.

I do not need accents on top of ◊ and □, so for now this commit removes
\vphantom until I find an alternative that works with tikz graphs.

Maybe I should argue that fixing the height of these symbols is outside
the scope of this package and should in any case be done be redefining
\Diamond and \Box.